refactor: adjust the project structure
