+++ apps/blog/src/components/mdx/hr.astro
@@ -1,6 +1,6 @@
 ---
 import { Separator } from "@/components/ui/separator";
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils";
 import type { SeparatorProps } from "@radix-ui/react-separator";
 import type { HTMLAttributes } from "astro/types";
 
@@ -14,3 +14,4 @@ type Props = Omit<SeparatorProps, "orientation"> & HTMLAttributes<"div">;
   className={cn("my-14", className)}
   {...props}
 />
+@/utils
