+++ apps/blog/src/components/mdx/spoiler.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils";
 
 export const Spoiler: React.FC<
   {} & {} & React.HTMLAttributes<HTMLSpanElement>
