+++ apps/blog/src/components/site-footer.astro
@@ -1,7 +1,7 @@
 ---
 import Link from "@/components/ui/link.astro";
 import { Separator } from "@/components/ui/separator";
-import { siteConfig } from "@/config/site";
+import { siteConfig } from "@/config";
 
 const { author, footer } = siteConfig;
 ---
@@ -24,3 +24,4 @@ const { author, footer } = siteConfig;
     }
   </p>
 </footer>
+@/site @/config
