+++ apps/blog/tsconfig.json
@@ -3,10 +3,10 @@
   "extends": "astro/tsconfigs/strict",
   "compilerOptions": {
     "baseUrl": ".",
-    "paths": { "@/*": ["./src/*"], "@/config/*": ["./config/*"] },
+    "paths": { "@/*": ["./src/*"] },
     "jsx": "react-jsx",
     "jsxImportSource": "react"
   },
-  "include": ["src/**/*", "config/*.ts", "*.config.ts"],
+  "include": ["src/**/*", "*.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
